move store file
